Split out `inbound::policy::test_util`
`Store::for_test` is stashed in the tcp test, so it's awkward to
reference its type. This change adds a new `test_util`
