+++ linkerd/app/inbound/src/accept.rs
@@ -116,7 +116,7 @@ impl svc::Param<AllowPolicy> for Accept {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{policy::Store, test_util};
+    use crate::{policy::test_util::Store, test_util};
     use futures::future;
     use linkerd_app_core::{
         svc::{NewService, ServiceExt},
@@ -125,10 +125,12 @@ mod tests {
     use linkerd_proxy_server_policy::{Authentication, Authorization, Meta, ServerPolicy};
     use std::sync::Arc;
 
+    const PROXY_PORT: u16 = 999;
+
     #[tokio::test(flavor = "current_thread")]
-    async fn default_allow() {
+    async fn policy() {
         let (io, _) = io::duplex(1);
-        let policies = Store::for_test(std::iter::once((
+        let policies = Store::fixed_for_test(std::iter::once((
             1000,
             ServerPolicy {
                 protocol: linkerd_proxy_server_policy::Protocol::Opaque(Arc::new([
@@ -149,9 +151,14 @@ mod tests {
                 }),
             },
         )));
+
         inbound()
             .with_stack(new_ok())
-            .push_accept(999, policies, new_panic("direct stack must not be built"))
+            .push_accept(
+                PROXY_PORT,
+                policies,
+                new_panic("direct stack must not be built"),
+            )
             .into_inner()
             .new_service(Target(1000))
             .oneshot(io)
@@ -161,33 +168,13 @@ mod tests {
 
     #[tokio::test(flavor = "current_thread")]
     async fn direct() {
-        let policies = Store::for_test(std::iter::once((
-            1000,
-            ServerPolicy {
-                protocol: linkerd_proxy_server_policy::Protocol::Opaque(Arc::new([
-                    Authorization {
-                        authentication: Authentication::Unauthenticated,
-                        networks: vec![Default::default()],
-                        meta: Arc::new(Meta::Resource {
-                            group: "policy.linkerd.io".into(),
-                            kind: "serverauthorization".into(),
-                            name: "testsaz".into(),
-                        }),
-                    },
-                ])),
-                meta: Arc::new(Meta::Resource {
-                    group: "policy.linkerd.io".into(),
-                    kind: "server".into(),
-                    name: "testsrv".into(),
-                }),
-            },
-        )));
         let (io, _) = io::duplex(1);
+        let policies = Store::fixed_for_test(std::iter::empty());
         inbound()
             .with_stack(new_panic("detect stack must not be built"))
-            .push_accept(999, policies, new_ok())
+            .push_accept(PROXY_PORT, policies, new_ok())
             .into_inner()
-            .new_service(Target(999))
+            .new_service(Target(PROXY_PORT))
             .oneshot(io)
             .await
             .expect("should succeed");
