+++ linkerd/app/inbound/src/policy/tcp.rs
@@ -10,6 +10,7 @@ use linkerd_app_core::{
 };
 use linkerd_proxy_server_policy::{Protocol, ServerPolicy};
 use std::{future::Future, pin::Pin, task};
+
 #[cfg(test)]
 mod tests;
